feat: docking, UI performance, profiling

+++ crates/astrelis-text/src/renderer/hybrid.rs
@@ -484,7 +484,7 @@ impl FontRenderer {
                 };
 
                 // Calculate scale factor from base size to target size
-                let target_size = f32::from_bits(cache_key.font_size_bits as u32);
+                let target_size = f32::from_bits(cache_key.font_size_bits);
                 let size_scale = target_size / sdf_entry.base_size;
 
                 // Scale placement based on size ratio
